Move pointer flow collector into pointer_flow
